refactor sections specializations and doctors
