+++ lib/features/home/presentation/widgets/doctors_list/doctors_shimmer_loading.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:shimmer/shimmer.dart';
 
-import '../../../../../../core/helpers/spacing.dart';
-import '../../../../../../core/theming/colors.dart';
+import '../../../../../../../core/helpers/spacing.dart';
+import '../../../../../../../core/theming/colors.dart';
 
 class DoctorsShimmerLoading extends StatelessWidget {
   const DoctorsShimmerLoading({super.key});
