+++ lib/features/home/presentation/widgets/specializations_list/doctors_specializations_loading_shimmer.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:shimmer/shimmer.dart';
 
-import '../../../../../../core/helpers/spacing.dart';
-import '../../../../../../core/theming/colors.dart';
+import '../../../../../../../core/helpers/spacing.dart';
+import '../../../../../../../core/theming/colors.dart';
 
 class SpecialityShimmerLoading extends StatelessWidget {
   const SpecialityShimmerLoading({super.key});
