+++ lib/features/home/presentation/widgets/specializations_list/list_view_specializations.dart
