+++ lib/main_production.dart
@@ -26,9 +26,9 @@ checkUserIfLoggedOrNot() async {
   String? token = await SharedPrefHelper.getSecuredString(
     SharedPrefKeys.userToken,
   );
-  if(token.isNullOrEmpty()){
-    isLoggedInUser=false;
-  }else{
-    isLoggedInUser=true;
+  if (token.isNullOrEmpty()) {
+    isLoggedInUser = false;
+  } else {
+    isLoggedInUser = true;
   }
 }
